Add seqinfo to all GRange objects
Addition of this field confers knowledge of sequence length and species
to the GRange objects. This will be very convenient in debugging,
especially when I make the leap to multiple species.

This change has also turned up a small inconsistency. I am getting
warnings that some intervals fall outside the expected chromosome
bounds. There are only a few cases, and all are off-by-one. I do not
think this is caused by mismatch of index conventions (0 versus 1
based). Thousands of intervals end exactly at the end of the scaffold
(after +1 to offset Synder's 0-base concention), but only a few go past
it. So I think it is a rarer edge case. Maybe the result of an
inconsistency in Synder.