Complete stable pre-label loop
The target loop and related infrastructure is all now functional. I have
added visuals for every major input to the labeling engine (not yet
done).

I removed the fo.cds and fo.mrna tables in favor of a more general list
of features. This change will require modifications to currently
commented-out or unused pieces of code if they are ever resurrected.