Put $HOME back into config
I am currently using a config that is suppose to be read by two
different languages (R and Bash). I tried to do the following:

x="~/path/to/file"

This works fine with R, but Bash does not expand the '~'. So Bash needs
$HOME. I should probably find a better solution.

Another annoying issue, is Bash does not allow space around the '='. The
config file is suppose to be edited by the user, so I can imagine this
causing problems.