Change requirements for input GFF to NCBI standard
Now the 9th field of the GFF files is required to have the tag
ID=([^;]+) and optionally the tag Parent=([^;]+). This will create a
natural and generalizeable way to deal with models (all models for a
given gene will have a link to their parent). However, it places a
difficult constraint on input.