Adjust scores with robust regression
Previously I was penalizing large search spaces by subtracting log2(mn)
from the raw score (where m and n are the query length and search
length).

E = Amne^(BS)

Where E is the expected number of hits to the database, S is the raw
score, and A and B are constants. This is the form of the equation used
by BLAST. This can be represented as a linear equation

S = a + b log(mn)

a and b can be estimated with ordinary least squares. However, there are
likely to be outliers. So I use a form of L1 robust regression (minimize
mean absolute error, rather than mean square error). Now for adjusted
scores, I use the residuals:

S_adjusted = S_observed - a - b log(mn)

My previous approach implicitly assumed b=1. This would, I think,
over-penalize large search spaces.

NOTE: I blindly used a robust regression tool, robustreg::robustRegBS,
with no understanding of the algorithm it uses. I need to verify that it
is doing what I think it is doing.