Fully functional configuration system
Config files are automatically built and linked to the appropriate
places.

preconfig.sh is linked to the prologue folder, where the data
preparation scripts reside.

runcondif.R is linked to the report folder where the data analysis
is done.

All filenames in the configs use absolute path, bypassing the need for
long file paths in the local code.